+++ src/red-gold-auton.cpp
@@ -1,5 +1,5 @@
-#include "vex.h";
-#include "blue-gold-auton.h";
+#include "vex.h"
+#include "autons.h"
 using namespace vex;
 
 void redGoldAutonStart() {
@@ -8,7 +8,7 @@ void redGoldAutonStart() {
     movething(directionType::fwd);
   }
   
-  while (abs(Sonar.distance(distanceUnits::in) - 42.5) > 2.5){
+  while (abs((int) (Sonar.distance(distanceUnits::in) - 42.5)) > 2.5){
     turnLeft();
   }
